deps/v8/test/inspector/debugger/set-breakpoints-active-expected.txt

Tests for set-breakpoints-active

Running test: testDeactivatedBreakpointsAfterReconnect
Breakpoints activated.
Paused. (reason: other)
Debugger break executed.
Breakpoints deactivated.
Reconnected.
Debugger break executed.

Running test: testDeactivatedBreakpointsAfterDisableEnable
Breakpoints activated.
Paused. (reason: other)
Debugger break executed.
Breakpoints deactivated.
Disabled.
Enabled.
Debugger break executed.

Running test: testDeactivateBreakpointsWhileDisabled
Breakpoints activated.
Paused. (reason: other)
Debugger break executed.
Disabled.
Breakpoints deactivated.
Enabled.
Debugger break executed.

Running test: testActivateBreakpointsWhileDisabled
Breakpoints deactivated.
Debugger break executed.
Disabled.
Breakpoints activated.
Enabled.
Paused. (reason: other)
Debugger break executed.
